src/requirements.txt
Django==1.10.1
django-grappelli==2.8.2
django-nested-admin==3.0.10
mysqlclient==1.3.9
python-monkey-business==1.0.0
six==1.10.0
vboxapi==1.0
virtualenv==15.0.3
